+++ Pods/Target Support Files/EVReflection-1eb814ba/EVReflection-1eb814ba-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_EVReflection_1eb814ba : NSObject
-@end
-@implementation PodsDummy_EVReflection_1eb814ba
-@end